+++ cardano-api/src/Cardano/Api/Experimental.hs
@@ -73,5 +73,5 @@ import Cardano.Api.Experimental.Plutus.Internal.ScriptWitness
 import Cardano.Api.Experimental.Plutus.Internal.Shim.LegacyScripts
 import Cardano.Api.Experimental.Simple.Script
 import Cardano.Api.Experimental.Tx
-import Cardano.Api.Internal.Experimental.TxBody.Certificate
+import Cardano.Api.Experimental.Tx.Internal.Certificate
 import Cardano.Api.Tx.Internal.Fee (evaluateTransactionExecutionUnitsShelley)
